PythonMovie/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-surprise
pydantic
python-multipart